Fix bug checking type in scan_iter
Fix #108
